+++ CHANGELOG.md
@@ -10,9 +10,10 @@ Developer only seen features (such as internal functions) may or may not be reco
 
 Changelog List
 --------------
-### <a name="2167"></a>v2.1.6.7 (10/12/2014)
+### <a name="2167"></a>v2.1.7 ()
 ```
 + Add title to the Single Post Link
++ Add ability to expand breadcrumbs on showthread
 ```
 ### <a name="2166"></a>v2.1.6.6 (10/10/2014)
 ```